requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
chess==1.11.2
colorama==0.4.6
cryptography==43.0.1
distro==1.9.0
et-xmlfile==1.1.0
frozenlist==1.7.0
geographiclib==2.0
geopy==2.4.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
multidict==6.6.3
nest-asyncio==1.6.0
numpy==2.1.1
openai==0.28.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pdfkit==1.0.0
pdfminer.six==20231228
pdfplumber==0.11.4
pefile==2023.2.7
pillow==10.4.0
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pyodbc==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PySocks==1.7.1
python-bidi==0.6.6
python-chess==1.999
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==22.2
pytz==2024.2
pywin32-ctypes==0.2.3
RapidFuzz==3.13.0
reportlab==4.4.2
requests==2.32.3
selenium==4.25.0
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tabula-py==2.9.3
thefuzz==0.22.1
tqdm==4.67.1
trio==0.26.2
trio-websocket==0.11.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.1
xhtml2pdf
